Refactoring TradeWidget into OrderDetails component
